Added Dependence of worxco_manual_crop_styles to have the image field use the manual crops created there to help minimize the number of crops that will be created.
Set the Image field to use crop_default_square and crop_medium_square and set both to be required when created a staff member.

Switched the views to display either crop_default_square or crop_medium_square depending on what was originally used.

Got the admin configuration of Title and Menu Title to work.